+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## Unreleased Changes
+
+## v0.1.12
+- Add pipeline for bumping release version automatically and updating changelog versions.
+- Add end to end testing automations
 
 ## 0.1.11
 - Revert inline runner execution logic